URI p-1040 accepted

+++ URI_midterm/p-1040/src/Main.java
@@ -9,17 +9,17 @@ public class Main {
     // Main
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
-        double N1 = s.nextDouble();
-        double N2 = s.nextDouble();
-        double N3 = s.nextDouble();
-        double N4 = s.nextDouble();
-        double N5 = 0;
+        float N1 = s.nextFloat();
+        float N2 = s.nextFloat();
+        float N3 = s.nextFloat();
+        float N4 = s.nextFloat();
+        float N5 = 0;
 
-        double average = (N1*2 + N2*3 + N3*4 + N4*1)/10.0;
+        float average = (N1*2 + N2*3 + N3*4 + N4*1)/10;
 
         if(average>=5.0 && average<=6.9)
         {
-            N5 = s.nextDouble();
+            N5 = s.nextFloat();
         }
 
         System.out.printf("Media: %.1f\n", average);
